Atualiza a versão do mod para 0.8.1 e traduz a descrição do mod para o inglês. Refatora a lógica de classes e experiência, movendo dados para o novo sistema de subclasses. Melhora a interface do usuário, desabilitando a UI antiga e implementando um novo sistema unificado. Ajusta a lógica de experiência para garantir que valores sejam inteiros e remove trechos de código obsoletos.

+++ Common/Systems/RPGActionSystem.cs
@@ -258,7 +258,7 @@ namespace Wolfgodrpg.Common.Systems
                 100f
             );
 
-            rpgPlayer.AddClassExperience("survivalist", 5f);
+            rpgPlayer.AddClassExperience("survivalist", (int)5f);
         }
     }
 }
